+++ src/HCOD.cpp
@@ -130,7 +130,7 @@ namespace soth
   {
     int r=0;
     for (size_t i=0; i<stages.size(); ++i)
-      r+= stages[i]->rank();
+      r+= (int) stages[i]->rank();
     return r;
   }
